OK-611 Lisätty testit LaskentaRunner-luokalle
